chore: decompose api in blueprints
